myagent/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.1
aiofiles==23.2.1
pydantic==2.6.3
python-jose==3.3.0
PyPDF2==3.0.1